Move playback overlay rows into composables
