+++ app/src/main/java/com/github/damontecres/wholphin/ui/playback/overlay/PlaybackController.kt
@@ -1,0 +1,110 @@
+package com.github.damontecres.wholphin.ui.playback.overlay
+
+import androidx.compose.foundation.focusable
+import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.focus.onFocusChanged
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import androidx.media3.common.Player
+import androidx.tv.material3.MaterialTheme
+import androidx.tv.material3.Text
+import com.github.damontecres.wholphin.R
+import com.github.damontecres.wholphin.data.model.BaseItem
+import com.github.damontecres.wholphin.ui.playback.ControllerViewState
+import com.github.damontecres.wholphin.ui.playback.PlaybackDialogType
+import org.jellyfin.sdk.model.api.MediaSegmentDto
+import kotlin.time.Duration
+
+@Composable
+fun PlaybackController(
+    item: BaseItem?,
+    nextState: OverlayViewState?,
+    playerControls: Player,
+    controllerViewState: ControllerViewState,
+    showPlay: Boolean,
+    showClock: Boolean,
+    previousEnabled: Boolean,
+    nextEnabled: Boolean,
+    seekEnabled: Boolean,
+    seekBack: Duration,
+    skipBackOnResume: Duration?,
+    seekForward: Duration,
+    onPlaybackActionClick: (PlaybackAction) -> Unit,
+    onClickPlaybackDialogType: (PlaybackDialogType) -> Unit,
+    onSeekBarChange: (Long) -> Unit,
+    currentSegment: MediaSegmentDto?,
+    onChangeState: (OverlayViewState) -> Unit,
+    modifier: Modifier = Modifier,
+    seekBarInteractionSource: MutableInteractionSource = remember { MutableInteractionSource() },
+) {
+    Column(
+        verticalArrangement = Arrangement.spacedBy(8.dp),
+        modifier = modifier,
+    ) {
+        Controller(
+            title = item?.title,
+            subtitle = item?.subtitleLong,
+            playerControls = playerControls,
+            controllerViewState = controllerViewState,
+            showPlay = showPlay,
+            showClock = showClock,
+            previousEnabled = previousEnabled,
+            nextEnabled = nextEnabled,
+            seekEnabled = seekEnabled,
+            seekBack = seekBack,
+            skipBackOnResume = skipBackOnResume,
+            seekForward = seekForward,
+            onPlaybackActionClick = onPlaybackActionClick,
+            onClickPlaybackDialogType = onClickPlaybackDialogType,
+            onSeekProgress = onSeekBarChange,
+            seekBarInteractionSource = seekBarInteractionSource,
+            nextState = nextState,
+            onNextStateFocus = {
+                nextState?.let { onChangeState.invoke(it) }
+            },
+            currentSegment = currentSegment,
+            modifier =
+            Modifier,
+            // Don't use key events because this control has vertical items so up/down is tough to manage
+        )
+        when (nextState) {
+            OverlayViewState.CHAPTERS -> {
+                Text(
+                    text = stringResource(R.string.chapters),
+                    style = MaterialTheme.typography.titleLarge,
+                    modifier =
+                        Modifier
+                            .padding(start = 16.dp, top = 0.dp)
+                            .onFocusChanged {
+                                if (it.isFocused) onChangeState.invoke(nextState)
+                            }.focusable(),
+                )
+            }
+
+            OverlayViewState.QUEUE -> {
+                Text(
+                    text = stringResource(R.string.queue),
+                    style = MaterialTheme.typography.titleLarge,
+                    modifier =
+                        Modifier
+                            .padding(start = 16.dp, top = 0.dp)
+                            .onFocusChanged {
+                                if (it.isFocused) onChangeState.invoke(nextState)
+                            }.focusable(),
+                )
+            }
+
+            else -> {
+                Spacer(Modifier.height(32.dp))
+            }
+        }
+    }
+}
